+++ controllers/courses.js
@@ -1,4 +1,5 @@
 import Course from "../models/Course.js";
+import Bootcamp from "../models/Bootcamp.js";
 import { parseQuery } from "../utils/parseQuery.js";
 
 export const getCourses = (req, res, next) => {
@@ -53,7 +54,28 @@ export const getCourse = (req, res, next) => {
 
 export const createCourse = (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId;
-  Course.create(req.body)
+  req.body.user = req.user.id;
+
+  Bootcamp.findById(req.params.bootcampId)
+    .then((bootcamp) => {
+      if (!bootcamp) {
+        return res
+          .status(400)
+          .json({ success: false, msg: "bootcamp doesn't exist" });
+      }
+
+      if (
+        bootcamp.user.toString() !== req.user.id &&
+        req.user.role !== "admin"
+      ) {
+        return res.status(401).json({
+          success: false,
+          msg: `User ${req.user.id} is not authorized to add a course to bootcamp ${req.params.bootcampId}`,
+        });
+      }
+
+      return Course.create(req.body);
+    })
     .then((course) => {
       res
         .status(201)
@@ -65,15 +87,27 @@ export const createCourse = (req, res, next) => {
 };
 
 export const updateCourse = (req, res, next) => {
-  console.log(req.body);
-  Course.findByIdAndUpdate(req.params.id, req.body)
+  Course.findById(req.params.id)
     .then((course) => {
       if (!course) {
         return res
           .status(400)
           .json({ success: false, msg: "course doesn't exist" });
       }
-      res.status(200).json({ success: true, data: course });
+
+      if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return res.status(401).json({
+          success: false,
+          msg: `User ${req.user.id} is not authorized to update course ${course._id}`,
+        });
+      }
+
+      return Course.findByIdAndUpdate(req.params.id, req.body);
+    })
+    .then((course) => {
+      res
+        .status(200)
+        .json({ success: true, data: course, msg: "course updated!!" });
     })
     .catch((err) => {
       next(err);
@@ -81,14 +115,27 @@ export const updateCourse = (req, res, next) => {
 };
 
 export const deleteCourse = (req, res, next) => {
-  Course.findByIdAndDelete(req.params.id)
+  Course.findById(req.params.id)
     .then((course) => {
       if (!course) {
         return res
           .status(400)
           .json({ success: false, msg: "course doesn't exist" });
       }
-      res.status(200).json({ success: true, data: {} });
+
+      if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return res.status(401).json({
+          success: false,
+          msg: `User ${req.user.id} is not authorized to delete course ${course._id}`,
+        });
+      }
+
+      return Course.findByIdAndDelete(req.params.id);
+    })
+    .then((course) => {
+      res
+        .status(200)
+        .json({ success: true, data: {}, msg: "course deleted!!" });
     })
     .catch((err) => {
       next(err);
